+++ tests/php/Unit/Run/Domain/Test/TestCommandOptionsTest.php
@@ -13,13 +13,41 @@ final class TestCommandOptionsTest extends TestCase
     {
         $options = TestCommandOptions::empty();
 
-        self::assertSame('{:filter nil}', $options->asPhelHashMap());
+        self::assertSame('{:filter nil :testdox false}', $options->asPhelHashMap());
     }
 
     public function test_random_filter(): void
     {
         $options = TestCommandOptions::fromArray(['filter' => 'example']);
 
-        self::assertSame('{:filter "example"}', $options->asPhelHashMap());
+        self::assertSame('{:filter "example" :testdox false}', $options->asPhelHashMap());
+    }
+
+    public function test_no_testdox(): void
+    {
+        $options = TestCommandOptions::fromArray([]);
+
+        self::assertSame('{:filter nil :testdox false}', $options->asPhelHashMap());
+    }
+
+    public function test_false_testdox(): void
+    {
+        $options = TestCommandOptions::fromArray(['testdox' => false]);
+
+        self::assertSame('{:filter nil :testdox false}', $options->asPhelHashMap());
+    }
+
+    public function test_null_testdox(): void
+    {
+        $options = TestCommandOptions::fromArray(['testdox' => null]);
+
+        self::assertSame('{:filter nil :testdox false}', $options->asPhelHashMap());
+    }
+
+    public function test_true_testdox(): void
+    {
+        $options = TestCommandOptions::fromArray(['testdox' => 'true']);
+
+        self::assertSame('{:filter nil :testdox true}', $options->asPhelHashMap());
     }
 }